Rename `Editor` component.
